저장

+++ src/main/java/org/example/kakaotechcampus_assignment2/domain/plan/service/DefaultPlanService.java
@@ -9,6 +9,8 @@ import org.example.kakaotechcampus_assignment2.domain.plan.dto.PlanResponseDto;
 import org.example.kakaotechcampus_assignment2.domain.plan.dto.PlanUpdateRequestDto;
 import org.example.kakaotechcampus_assignment2.domain.plan.entity.AuthorInfo;
 import org.example.kakaotechcampus_assignment2.domain.plan.entity.Plan;
+import org.example.kakaotechcampus_assignment2.domain.plan.exception.InvalidPlanPasswordException;
+import org.example.kakaotechcampus_assignment2.domain.plan.exception.PlanNotFoundException;
 import org.example.kakaotechcampus_assignment2.domain.plan.repository.PlanRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -17,7 +19,6 @@ import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.Collections;
 import java.util.List;
-import java.util.Objects;
 
 @Service
 @RequiredArgsConstructor
@@ -48,26 +49,23 @@ public class DefaultPlanService implements PlanService {
 
     @Override
     public PlanResponseDto getPlanById(Long id) {
-        return planRepository.findById(id);
+        PlanResponseDto plan = planRepository.findById(id);
+        if (plan == null) {
+            throw new PlanNotFoundException("일정 정보를 찾을 수 없습니다.");
+        }
+        return plan;
     }
 
     @Transactional
     @Override
     public PlanResponseDto updatePlan(Long planId, PlanUpdateRequestDto request) {
         Plan existingPlan = planRepository.findPlanEntityById(planId);
-
         if (existingPlan == null) {
-            return null;
+            throw new PlanNotFoundException("수정할 일정 정보를 찾을 수 없습니다.");
         }
-
-        if (!Objects.equals(existingPlan.authorInfo().name(), request.memberId())) {
-            return null;
+        if (!existingPlan.pwd().equals(request.pwd())) {
+            throw new InvalidPlanPasswordException("비밀번호가 일치하지 않습니다.");
         }
-
-        if (!Objects.equals(existingPlan.pwd(), request.pwd())) {
-            return null;
-        }
-
         Plan planToUpdate = new Plan(
                 existingPlan.id(),
                 existingPlan.authorInfo(),
@@ -75,15 +73,20 @@ public class DefaultPlanService implements PlanService {
                 existingPlan.pwd(),
                 existingPlan.createdAt(),
                 LocalDateTime.now());
-
         planRepository.update(planId, planToUpdate);
-
         return planRepository.findById(planId);
     }
 
     @Transactional
     @Override
     public void deletePlan(Long planId, String pwd, Long memberId) {
+        Plan existingPlan = planRepository.findPlanEntityById(planId);
+        if (existingPlan == null) {
+            throw new PlanNotFoundException("삭제할 일정 정보를 찾을 수 없습니다.");
+        }
+        if (!existingPlan.pwd().equals(pwd)) {
+            throw new InvalidPlanPasswordException("비밀번호가 일치하지 않습니다.");
+        }
         planRepository.deleteById(planId);
     }
